HockeyAPI/requirements.txt
requests~=2.32.5
beautifulsoup4~=4.14.3
Flask~=3.1.2
ics~=0.7.2
pdfplumber~=0.11.8
python-dateutil~=2.9.0.post0
gunicorn~=21.2.0
Flask-Caching~=2.3.1